requirements.txt
Flask
Flask-WTF
Flask-Flash
WTForms
ReportLab
python-docx
